policy/test_global.te: support running in a selinux namespace
To run the testsuite in a child selinux namespace, we need to allow
unconfined_t (or whatever domain creates the child namespace) in the
parent namespace all the permissions used by the test domains in the
child namespace. Otherwise, the tests fail due to failing permission
checks against the parent namespace using the parent cred.

Signed-off-by: Stephen Smalley <stephen.smalley.work@gmail.com>