tests/capable_file: fix fsetid test
The current test for the SELinux fsetid capability check tries to
set the setgid bit on a file with a different group ownership and
compares the file mode before and after to determine success/failure.
However, for reasons unknown, the current test runs chmod g+rs
which tries to set both the read bit and setgid bit, and will change
the read bit if not already set, causing the test to fail if the umask
doesn't allow group read by default. Fix the test to only run chmod g+s.

Signed-off-by: Stephen Smalley <stephen.smalley.work@gmail.com>