tests/mac_admin: disable test of getting undefined label on nfs
We already had a workaround in the test script to expect a different
result on NFS to the unique aspects of labeled NFS, but this
workaround won't work when/if SELinux namespaces are enabled
(due to the introduction of the global SID table and its
correct preservation of the original context value) and was
really only papering over the lack of support. Just disable
this particular test when run on NFS filesystems.

Signed-off-by: Stephen Smalley <stephen.smalley.work@gmail.com>