tests/{inet_socket,sctp}/test: remove use of /sys/fs/selinux/unshare
Remove use of /sys/fs/selinux/unshare due to planned deprecation.
Note that we cannot merely substitute running the selinuxunshared
program or otherwise using the is_selinux_unshared() libselinux
API because it returns 0 again after a policy has been loaded.
If/when we decide we truly need a way to detect that we are in
a fully initialized non-init SELinux namespace, revisit that as
a new interface. In the interim, just always set is_unshared to
1 on this branch and only use it when testing child namespaces.

Signed-off-by: Stephen Smalley <stephen.smalley.work@gmail.com>